+++ Cargo.toml
@@ -24,6 +24,7 @@ members = [
     "crates/perl-lsp-providers",
     "crates/perl-lsp-rename",
     "crates/perl-lsp-protocol",
+    "crates/perl-lsp-request-params",
     "crates/perl-lsp-transport",
     "crates/perl-lsp-tooling",
     "crates/perl-subprocess-runtime",
@@ -186,6 +187,7 @@ allow = [
     "perl-lsp-launcher",
     "perl-lsp-limits",
     "perl-lsp-protocol",
+    "perl-lsp-request-params",
     "perl-lsp-transport",
 
     # Tier 6 — Module resolution higher tiers
@@ -255,6 +257,7 @@ perl-text-line = { path = "crates/perl-text-line", version = "0.10.0" }
 perl-content-length-framing = { path = "crates/perl-content-length-framing", version = "0.10.0" }
 perl-subprocess-runtime = { path = "crates/perl-subprocess-runtime", version = "0.10.0" }
 perl-lsp-protocol = { path = "crates/perl-lsp-protocol", version = "0.10.0" }
+perl-lsp-request-params = { path = "crates/perl-lsp-request-params", version = "0.10.0" }
 perl-uri = { path = "crates/perl-uri", version = "0.10.0" }
 perl-path-security = { path = "crates/perl-path-security", version = "0.10.0" }
 perl-workspace-discovery = { path = "crates/perl-workspace-discovery", version = "0.10.0" }
