+++ crates/perl-lsp-request-params/README.md
@@ -1,0 +1,4 @@
+# perl-lsp-request-params
+
+Reusable helpers for extracting strongly-typed LSP request parameters from
+JSON payloads.
