+++ crates/perl-lsp/Cargo.toml
@@ -31,6 +31,7 @@ perl-lsp-limits = { workspace = true }
 perl-lsp-feature-governance = { workspace = true }
 perl-lsp-tooling = { workspace = true }
 perl-lsp-protocol = { workspace = true }
+perl-lsp-request-params = { workspace = true }
 perl-lsp-transport = { workspace = true }
 perl-lsp-code-actions = { workspace = true }
 perl-lsp-completion = { workspace = true }
